Refactor: FireStorage <> UseCase 의존성 제거 및 UseCase 통합
