+++ Kabinett/Domain/UseCase/LetterWriteUseCase/LetterWriteUseCase.swift
@@ -27,4 +27,8 @@ protocol LetterWriteUseCase {
     ) async -> Result<Bool, any Error>
     func findWriter(by query: String) async -> [Writer]
     func getCurrentWriter() -> AnyPublisher<Writer, Never>
+    
+    func loadEnvelopes() async -> Result<[String], any Error>
+    func loadStamps() async -> Result<[String], any Error>
+    func loadStationeries() async -> Result<[String], any Error>
 }
