+++ Kabinett/Presentation/ViewModel/LetterWrite/StationerySelectionViewModel.swift
@@ -13,9 +13,9 @@ class StationerySelectionViewModel: ObservableObject {
     @Published var selectedIndex: (Int, Int) = (0, 0)
     @Published var stationerys: [String] = []
     
-    private let useCase: LetterWriteLoadStuffUseCase
+    private let useCase: LetterWriteUseCase
     
-    init(useCase: LetterWriteLoadStuffUseCase) {
+    init(useCase: LetterWriteUseCase) {
         self.useCase = useCase
         Task {
             await loadStationeries()
