+++ Kabinett/Service/Letter/DefaultNormalLetterUseCase.swift
@@ -117,4 +117,19 @@ extension DefaultNormalLetterUseCase: LetterWriteUseCase {
             .compactMap { $0 }
             .eraseToAnyPublisher()
     }
+    
+    // 편지봉투 로딩
+    func loadEnvelopes() async -> Result<[String], any Error> {
+        await letterStorageManager.loadStorage(path: "Envelopes")
+    }
+    
+    // 우표 로딩
+    func loadStamps() async -> Result<[String], any Error> {
+        await letterStorageManager.loadStorage(path: "Stamps")
+    }
+    
+    // 편지지 로딩
+    func loadStationeries() async -> Result<[String], any Error> {
+        await letterStorageManager.loadStorage(path: "Stationeries")
+    }
 }
